requirements.txt
certifi==2017.4.17
chardet==3.0.3
idna==2.5
plaid-python==2.0.2
PyJWT==1.5.0
PySocks==1.6.7
pytz==2017.2
requests==2.17.3
six==1.10.0
twilio==6.3.0
urllib3==1.21.1
